Refactore test resolver strategy
