+++ src/main/java/com/askimed/nf/test/core/TestExecutionEngine.java
@@ -22,7 +22,7 @@ import groovy.json.JsonOutput;
 
 public class TestExecutionEngine {
 
-	private List<File> scripts;
+	private List<ITestSuite> testSuits;
 
 	private ITestExecutionListener listener = new AnsiTestExecutionListener();
 
@@ -40,19 +40,9 @@ public class TestExecutionEngine {
 
 	private boolean cleanSnapshot = false;
 
-	private String libDir = "";
-
-	private PluginManager pluginManager = null;
-
-	private TagQuery tagQuery = new TagQuery();
-
 	private boolean dryRun = false;
 
 	private static Logger log = LoggerFactory.getLogger(TestExecutionEngine.class);
-
-	public void setScripts(List<File> scripts) {
-		this.scripts = scripts;
-	}
 
 	public void setDebug(boolean debug) {
 		this.debug = debug;
@@ -87,64 +77,12 @@ public class TestExecutionEngine {
 		this.cleanSnapshot = cleanSnapshot;
 	}
 
-	public void setTagQuery(TagQuery tagQuery) {
-		this.tagQuery = tagQuery;
-	}
-
-	public void setLibDir(String libDir) {
-		this.libDir = libDir;
+	public void setDryRun(boolean dryRun) {
+		this.dryRun = dryRun;
 	}
 
 	public void setListener(ITestExecutionListener listener) {
 		this.listener = listener;
-	}
-
-	public void setPluginManager(PluginManager pluginManager) {
-		this.pluginManager = pluginManager;
-	}
-
-	protected List<ITestSuite> parse(TagQuery tagQuery) throws Throwable {
-
-		List<ITestSuite> testSuits = new Vector<ITestSuite>();
-
-		for (File script : scripts) {
-			String testId = null;
-			if (script.getAbsolutePath().contains("@")) {
-				String[] tiles = script.getAbsolutePath().split("@");
-				script = new File(tiles[0]);
-				testId = tiles[1];
-			}
-			if (!script.exists()) {
-				throw new Exception("Test file '" + script.getAbsolutePath() + "' not found.");
-			}
-			ITestSuite testSuite = TestSuiteBuilder.parse(script, libDir, pluginManager);
-
-			boolean empty = true;
-
-			for (ITest test : testSuite.getTests()) {
-				if (testId != null) {
-					if (!test.getHash().startsWith(testId)) {
-						test.skip();
-					}
-				}
-
-				if (!tagQuery.matches(test)) {
-					test.skip();
-				}
-
-				if (!test.isSkipped()) {
-					empty = false;
-				}
-
-			}
-
-			if (!empty) {
-				testSuits.add(testSuite);
-			}
-		}
-
-		return testSuits;
-
 	}
 
 	public int execute() throws Throwable {
@@ -158,9 +96,7 @@ public class TestExecutionEngine {
 			}
 		}
 
-		List<ITestSuite> testSuits = parse(tagQuery);
-
-		if (testSuits.size() == 0) {
+		if (testSuits.isEmpty()) {
 			System.out.println(AnsiColors.red("Error: no valid tests found."));
 			System.out.println();
 			return 1;
@@ -261,165 +197,7 @@ public class TestExecutionEngine {
 
 	}
 
-	public int listTests(OutputFormat format) throws Throwable {
-
-		if (configFile != null) {
-			if (!configFile.exists()) {
-				System.out.println(
-						AnsiColors.red("Error: Test config file '" + configFile.getAbsolutePath() + "'not found"));
-				System.out.println();
-				return 1;
-			}
-		}
-
-		List<ITestSuite> testSuits = parse(tagQuery);
-
-		if (testSuits.size() == 0) {
-			System.out.println(AnsiColors.red("Error: no valid tests found."));
-			System.out.println();
-			return 1;
-		}
-
-		switch (format) {
-		case JSON:
-		case json:
-			printTestsAsJson(testSuits);
-			break;
-		case RAW:
-		case raw:
-			printTestsAsList(testSuits);
-			break;
-		case CSV:
-		case csv:
-			printTestsAsCsv(testSuits);
-			break;
-		default:
-			printTestsPretty(testSuits);
-			break;
-		}
-
-		return 0;
-
-	}
-
-	public int listTags(OutputFormat format) throws Throwable {
-
-		if (configFile != null) {
-			if (!configFile.exists()) {
-				System.out.println(
-						AnsiColors.red("Error: Test config file '" + configFile.getAbsolutePath() + "'not found"));
-				System.out.println();
-				return 1;
-			}
-		}
-
-		List<ITestSuite> testSuits = parse(tagQuery);
-
-		if (testSuits.size() == 0) {
-			System.out.println(AnsiColors.red("Error: no valid tests found."));
-			System.out.println();
-			return 1;
-		}
-
-		Set<String> tags = new HashSet<String>();
-		for (ITestSuite testSuite : testSuits) {
-			tags.addAll(testSuite.getTags());
-			for (ITest test : testSuite.getTests()) {
-				tags.addAll(test.getTags());
-			}
-		}
-
-		switch (format) {
-		case JSON:
-		case json:
-			printTagsAsJson(tags);
-			break;
-		case CSV:
-		case csv:
-			printTagsAsCsv(tags);
-			break;
-		default:
-			printTagsPretty(tags);
-			break;
-		}
-
-		return 0;
-
-	}
-
-	private void printTestsAsJson(List<ITestSuite> testSuits) {
-		List<String> tests = new Vector<String>();
-		int index = 0;
-		for (ITestSuite testSuite : testSuits) {
-			for (ITest test : testSuite.getTests()) {
-				tests.add(scripts.get(index).getAbsolutePath() + "@" + test.getHash().substring(0, 8));
-			}
-			index++;
-		}
-		System.out.println(JsonOutput.toJson(tests));
-	}
-
-	private void printTestsAsList(List<ITestSuite> testSuits) {
-		int index = 0;
-		for (ITestSuite testSuite : testSuits) {
-			for (ITest test : testSuite.getTests()) {
-				System.out.println(scripts.get(index).getAbsolutePath() + "@" + test.getHash().substring(0, 8));
-			}
-			index++;
-		}
-	}
-
-	private void printTestsAsCsv(List<ITestSuite> testSuits) {
-		List<String> tests = new Vector<String>();
-		int index = 0;
-		for (ITestSuite testSuite : testSuits) {
-			for (ITest test : testSuite.getTests()) {
-				tests.add(scripts.get(index).getAbsolutePath() + "@" + test.getHash().substring(0, 8));
-			}
-			index++;
-		}
-		System.out.println(String.join(",", tests));
-	}
-
-	private void printTestsPretty(List<ITestSuite> testSuits) {
-		int index = 0;
-		int count = 0;
-		for (ITestSuite testSuite : testSuits) {
-
-			System.out.println();
-			System.out.println("[" + FileUtil.makeRelative(baseDir, scripts.get(index)) + "] "
-					+ AnsiText.bold(testSuite.getName()));
-			System.out.println();
-
-			for (ITest test : testSuite.getTests()) {
-				System.out.println(AnsiText.padding("[" + FileUtil.makeRelative(baseDir, scripts.get(index)) + "@"
-						+ test.getHash().substring(0, 8) + "] " + AnsiText.bold(test.getName()), 2));
-				count++;
-
-			}
-			index++;
-		}
-
-		System.out.println();
-		System.out.println("Found " + count + " tests.");
-		System.out.println();
-	}
-
-	private void printTagsAsJson(Set<String> tags) {
-		System.out.println(JsonOutput.toJson(tags));
-	}
-
-	private void printTagsAsCsv(Set<String> tags) {
-		System.out.println(String.join(",", tags));
-	}
-
-	private void printTagsPretty(Set<String> tags) {
-		for (String tag : tags) {
-			System.out.println(tag);
-		}
-	}
-
-	public void setDryRun(boolean dryRun) {
-		this.dryRun = dryRun;
+	public void setTestSuites(List<ITestSuite> testSuits) {
+		this.testSuits = testSuits;
 	}
 }
